refactor: unify cancellation reporting behind reportError
