+++ app/views/ChangeAvatarView/index.tsx
@@ -9,7 +9,7 @@ import { textInputDebounceTime } from '../../lib/constants/debounceConfig';
 import KeyboardView from '../../containers/KeyboardView';
 import sharedStyles from '../Styles';
 import scrollPersistTaps from '../../lib/methods/helpers/scrollPersistTaps';
-import { showConfirmationAlert, showErrorAlert } from '../../lib/methods/helpers/info';
+import { reportError, showConfirmationAlert } from '../../lib/methods/helpers/info';
 import { useTheme } from '../../theme';
 import SafeAreaView from '../../containers/SafeAreaView';
 import * as List from '../../containers/List';
@@ -29,7 +29,6 @@ import ImagePicker, { type Image } from '../../lib/methods/helpers/ImagePicker/I
 import { compareServerVersion, isImageURL, useDebounce } from '../../lib/methods/helpers';
 import { ControlledFormTextInput } from '../../containers/TextInput';
 import { HeaderBackButton } from '../../containers/Header/components/HeaderBackButton';
-import { isTwoFactorCancelled } from '../../lib/services/twoFactor';
 
 enum AvatarStateActions {
 	CHANGE_AVATAR = 'CHANGE_AVATAR',
@@ -173,11 +172,8 @@ const ChangeAvatarView = () => {
 			}
 			isDirty.current = false;
 		} catch (e: any) {
-			if (isTwoFactorCancelled(e)) {
-				return;
-			}
-			log(e);
-			return showErrorAlert(e.message, I18n.t('Oops'));
+			reportError(e, e.message, I18n.t('Oops'));
+			return;
 		} finally {
 			setSaving(false);
 		}
