+++ app/views/E2EEncryptionSecurityView/ChangePassword.tsx
@@ -4,12 +4,11 @@ import Clipboard from '@react-native-clipboard/clipboard';
 
 import { useTheme } from '../../theme';
 import I18n from '../../i18n';
-import log, { events, logEvent } from '../../lib/methods/helpers/log';
+import { events, logEvent } from '../../lib/methods/helpers/log';
 import { FormTextInput } from '../../containers/TextInput';
 import Button from '../../containers/Button';
 import { Encryption } from '../../lib/encryption';
-import { isTwoFactorCancelled } from '../../lib/services/twoFactor';
-import { showConfirmationAlert, showErrorAlert } from '../../lib/methods/helpers/info';
+import { reportError, showConfirmationAlert } from '../../lib/methods/helpers/info';
 import EventEmitter from '../../lib/methods/helpers/events';
 import { LISTENER } from '../../containers/Toast';
 import { useAppSelector } from '../../lib/hooks/useAppSelector';
@@ -49,11 +48,7 @@ const ChangePassword = () => {
 					newPasswordInputRef?.current?.clear();
 					newPasswordInputRef?.current?.blur();
 				} catch (e) {
-					if (isTwoFactorCancelled(e)) {
-						return;
-					}
-					log(e);
-					showErrorAlert(I18n.t('E2E_encryption_change_password_error'));
+					reportError(e, I18n.t('E2E_encryption_change_password_error'));
 				}
 			}
 		});
