+++ app/views/E2EEncryptionSecurityView/index.tsx
@@ -6,14 +6,13 @@ import { useNavigation } from '@react-navigation/native';
 
 import * as List from '../../containers/List';
 import I18n from '../../i18n';
-import log, { events, logEvent } from '../../lib/methods/helpers/log';
+import { events, logEvent } from '../../lib/methods/helpers/log';
 import { useTheme } from '../../theme';
 import SafeAreaView from '../../containers/SafeAreaView';
 import Button from '../../containers/Button';
 import { logout } from '../../actions/login';
-import { showConfirmationAlert, showErrorAlert } from '../../lib/methods/helpers/info';
+import { reportError, showConfirmationAlert } from '../../lib/methods/helpers/info';
 import { e2eResetOwnKey } from '../../lib/services/restApi';
-import { isTwoFactorCancelled } from '../../lib/services/twoFactor';
 import { type SettingsStackParamList } from '../../stacks/types';
 import ChangePassword from './ChangePassword';
 import { styles } from './styles';
@@ -43,11 +42,7 @@ const E2EEncryptionSecurityView = () => {
 						dispatch(logout());
 					}
 				} catch (e) {
-					if (isTwoFactorCancelled(e)) {
-						return;
-					}
-					log(e);
-					showErrorAlert(I18n.t('E2E_encryption_reset_error'));
+					reportError(e, I18n.t('E2E_encryption_reset_error'));
 				}
 			}
 		});
