+++ app/views/ProfileView/index.tsx
@@ -26,7 +26,8 @@ import EventEmitter from '../../lib/methods/helpers/events';
 import { events, logEvent } from '../../lib/methods/helpers/log';
 import scrollPersistTaps from '../../lib/methods/helpers/scrollPersistTaps';
 import { saveUserProfile } from '../../lib/services/restApi';
-import { twoFactor, isTwoFactorCancelled } from '../../lib/services/twoFactor';
+import { twoFactor } from '../../lib/services/twoFactor';
+import { isTwoFactorCancelled } from '../../lib/services/twoFactorCancelled';
 import { getUserSelector } from '../../selectors/login';
 import { type ProfileStackParamList } from '../../stacks/types';
 import { useTheme } from '../../theme';
